# sql/parser: Address performance concerns with default DECIMAL type