Addresses performance concerns brought up in https://github.com/cockroachdb/cockroach/pull/6752.

#### Changes:
- make sure decimal references do not allocate on evaluation
- add a temporary working decimal to `EvalContext`

  *(this could be a lot cleaner if we passed around a pointer to `EvalContext`)*
- return statically allocated errors from `asInt64`

#### Perf impact:
```
name                 old time/op    new time/op    delta
Select3_Cockroach-4    1.33ms ± 4%    1.08ms ± 3%  -18.41%  (p=0.000 n=20+19)

name                 old alloc/op   new alloc/op   delta
Select3_Cockroach-4     182kB ± 0%     139kB ± 0%  -23.70%  (p=0.000 n=18+20)

name                 old allocs/op  new allocs/op  delta
Select3_Cockroach-4     3.19k ± 0%     2.26k ± 0%  -28.90%  (p=0.000 n=17+20)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/6787)
<!-- Reviewable:end -->