@petermattis 2016-05-19
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

<details><summary><i>Previously, nvanbenschoten (Nathan VanBenschoten) wrote…</i></summary>
> #### sql/parser: Address performance concerns with default DECIMAL type
> 
> Addresses performance concerns brought up in https://github.com/cockroachdb/cockroach/pull/6752.> 
> > 
> #### Changes:> 
> - make sure decimal references do not allocate on evaluation> 
> - add a temporary working decimal to `EvalContext`> 
> > 
>   *(this could be a lot cleaner if we passed around a pointer to `EvalContext`)*> 
> - return statically allocated errors from `asInt64`> 
> > 
> #### Perf impact:> 
> ```> 
> name                 old time/op    new time/op    delta> 
> Select3_Cockroach-4    1.33ms ± 4%    1.08ms ± 3%  -18.41%  (p=0.000 n=20+19)> 
> > 
> name                 old alloc/op   new alloc/op   delta> 
> Select3_Cockroach-4     182kB ± 0%     139kB ± 0%  -23.70%  (p=0.000 n=18+20)> 
> > 
> name                 old allocs/op  new allocs/op  delta> 
> Select3_Cockroach-4     3.19k ± 0%     2.26k ± 0%  -28.90%  (p=0.000 n=17+20)> 
> ```
> 
> 
> 
</details>


---

Reviewed 2 of 2 files at r1, 3 of 3 files at r2.
Review status: 4 of 5 files reviewed at latest revision, 2 unresolved discussions, some commit checks failed.

---

*[sql/parser/constant.go, line 145 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6787#-KI9ARbi1EzSsd9_dcRU:-KI9ARbjfpHoUlOrApLy:-1524701312) ([raw file](https://github.com/cockroachdb/cockroach/blob/f8e672425f6f75771d3a8a6abfa0c2a462361028/sql/parser/constant.go#L145)):*
> ```Go
> // }
> 
> // These errors must be statically allocated, because they are returned in the
> ```

`s/must be/are/g`

---

*[sql/parser/eval.go, line 1252 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6787#-KI9AGSADgTOkhjSeNMZ:-KI9AGSB1Hm4Ul7Ce8Lo:1382888055) ([raw file](https://github.com/cockroachdb/cockroach/blob/85b46ade2927262bddd9b35b80f7acad7ea4d522/sql/parser/eval.go#L1252)):*
> ```Go
> 		return ctx.TmpDec
> 	}
> 	return new(inf.Dec)
> ```

Seems like `EvalContext` should be a pointer so that you could allocate `TmpDec` on the first access. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6787#-:-KI9AZmSm1yHJoG83aUO:1417088721)*
<!-- Sent from Reviewable.io -->