@nvanbenschoten 2016-05-19
TFTR.

<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary>
> :lgtm:
</details>


---

Review status: 4 of 5 files reviewed at latest revision, 2 unresolved discussions, some commit checks pending.

---

*[sql/parser/constant.go, line 145 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6787#-KI9ARbi1EzSsd9_dcRU:-KI9_KEO6GYicJqDPAH2:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/f8e672425f6f75771d3a8a6abfa0c2a462361028/sql/parser/constant.go#L145)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary>
> `s/must be/are/g`
</details>
Done.

---

*[sql/parser/eval.go, line 1252 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6787#-KI9AGSADgTOkhjSeNMZ:-KI9_KRu8CsDd9lV1zfL:948355439) ([raw file](https://github.com/cockroachdb/cockroach/blob/85b46ade2927262bddd9b35b80f7acad7ea4d522/sql/parser/eval.go#L1252)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary>
> Seems like `EvalContext` should be a pointer so that you could allocate `TmpDec` on the first access. 
</details>
I agree (see the PR message) and lazily allocating the `tmpDec` is exactly what I was planning to do. However, when I began looking into making the change, it looked like a pretty substantial refactor. I didn't want to mix that into this change, but I'll open an issue. https://github.com/cockroachdb/cockroach/issues/6809

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6787#-:-KI9bvQEYM_fJI7uP63R:-485616679)*
<!-- Sent from Reviewable.io -->